requirements/dev.txt
-r base.txt

# Debug
django-debug-toolbar==4.4.*
ipython==8.*

# Testing
pytest==8.*
pytest-django==4.9.*
pytest-cov==5.*
pytest-asyncio==0.24.*
factory-boy==3.3.*
faker==30.*
freezegun==1.5.*
responses==0.25.*
model-bakery==1.19.*

# Code Quality
ruff==0.7.*
black==24.*
isort==5.*
mypy==1.11.*
django-stubs==5.*
djangorestframework-stubs==3.15.*

# Security
bandit==1.7.*
safety==3.*
semgrep==1.*
